git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@16222 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/svg.c
@@ -2962,7 +2962,7 @@ static Image *ReadSVGImage(const ImageInfo *image_info,ExceptionInfo *exception)
             stride=4*image->columns;
 #if defined(MAGICKCORE_PANGOCAIRO_DELEGATE)
             stride=(size_t) cairo_format_stride_for_width(CAIRO_FORMAT_ARGB32,
-              image->columns);
+              (int) image->columns);
 #endif
             pixel_info=AcquireVirtualMemory(stride,image->rows*sizeof(*pixels));
             if (pixel_info == (MemoryInfo *) NULL)
@@ -2976,7 +2976,8 @@ static Image *ReadSVGImage(const ImageInfo *image_info,ExceptionInfo *exception)
             (void) SetImageBackgroundColor(image);
 #if defined(MAGICKCORE_CAIRO_DELEGATE)
             cairo_surface=cairo_image_surface_create_for_data(pixels,
-              CAIRO_FORMAT_ARGB32,image->columns,image->rows,4*image->columns);
+              CAIRO_FORMAT_ARGB32,(int) image->columns,(int) image->rows,
+              4*(int) image->columns);
             if (cairo_surface == (cairo_surface_t *) NULL)
               {
                 pixel_info=RelinquishVirtualMemory(pixel_info);
